+++ test/unit/AccountClient.test.ts
@@ -34,7 +34,7 @@ describe("AccountClient", () => {
         });
 
         it("httpClient initialized", () => {
-            expect(client.httpClient.client.defaults).not.to.eql(180000);
+            expect(client.httpClient.client).to.be.a("function");
         });
     });